+++ src/damage_system.rs
@@ -1,6 +1,6 @@
 use specs::prelude::*;
 use super::{
-    CombatStats, SufferDamage, Player, gamelog::GameLog, Name, RunState,
+    Pools, SufferDamage, Player, gamelog::GameLog, Name, RunState,
     Position, Map,
 };
 
@@ -8,7 +8,7 @@ pub struct DamageSystem {}
 
 impl<'a> System<'a> for DamageSystem {
     type SystemData = (
-        WriteStorage<'a, CombatStats>,
+        WriteStorage<'a, Pools>,
         WriteStorage<'a, SufferDamage>,
         ReadStorage<'a, Position>,
         WriteExpect<'a, Map>,
@@ -19,7 +19,7 @@ impl<'a> System<'a> for DamageSystem {
         let (mut stats, mut damage, positions, mut map, entities) = data;
 
         for (entity, mut stats, damage) in (&entities, &mut stats, &damage).join() {
-            stats.hp -= damage.amount.iter().sum::<i32>();
+            stats.hit_points.current -= damage.amount.iter().sum::<i32>();
             let pos = positions.get(entity);
             if let Some(pos) = pos {
                 let idx = map.xy_idx(pos.x, pos.y);
@@ -36,13 +36,13 @@ pub fn delete_the_dead(ecs: &mut World) {
 
     // Using a scope to make the borrow checker happy
     {
-        let combat_stats = ecs.read_storage::<CombatStats>();
+        let combat_stats = ecs.read_storage::<Pools>();
         let players = ecs.read_storage::<Player>();
         let names = ecs.read_storage::<Name>();
         let entities = ecs.entities();
         let mut log = ecs.write_resource::<GameLog>();
         for (entity, stats) in (&entities, &combat_stats).join() {
-            if stats.hp < 1 {
+            if stats.hit_points.current < 1 {
                 let player = players.get(entity);
                 match player {
                     None => {