+++ src/main.rs
@@ -44,6 +44,8 @@ mod particle_system;
 mod hunger_system;
 mod trigger_system;
 pub mod map_builders;
+mod gamesystem;
+pub use gamesystem::*;
 
 // Constants
 const SHOW_MAPGEN_VISUALISER: bool = true;
@@ -368,12 +370,7 @@ impl State {
         // Notify the player and regenerate some health.
         let player_entity = self.ecs.fetch::<Entity>();
         let mut gamelog = self.ecs.fetch_mut::<gamelog::GameLog>();
-        gamelog.entries.push("You descend to the next level, and take a moment to catch your breath.".to_string());
-        let mut player_health_store = self.ecs.write_storage::<CombatStats>();
-        let player_health = player_health_store.get_mut(*player_entity);
-        if let Some(player_health) = player_health {
-            player_health.hp = i32::max(player_health.hp, player_health.max_hp/2); // Heal up to half health
-        }
+        gamelog.entries.push("You descend to the next level.".to_string());
     }
 
     fn game_over_cleanup(&mut self) {
@@ -462,7 +459,6 @@ fn main() -> rltk::BError {
     gs.ecs.register::<Vendor>();
     gs.ecs.register::<Name>();
     gs.ecs.register::<BlocksTile>();
-    gs.ecs.register::<CombatStats>();
     gs.ecs.register::<WantsToMelee>();
     gs.ecs.register::<SufferDamage>();
     gs.ecs.register::<Item>();
@@ -495,6 +491,9 @@ fn main() -> rltk::BError {
     gs.ecs.register::<BlocksVisibility>();
     gs.ecs.register::<Door>();
     gs.ecs.register::<Quips>();
+    gs.ecs.register::<Attributes>();
+    gs.ecs.register::<Skills>();
+    gs.ecs.register::<Pools>();
 
     gs.ecs.insert(SimpleMarkerAllocator::<SerializeMe>::new());